+++ crates/pdk-api/Cargo.toml
@@ -8,6 +8,7 @@ homepage = "https://moonrepo.dev/proto"
 repository = "https://github.com/moonrepo/proto"
 
 [dependencies]
+system_packages = { version = "0.0.1", path = "../system-packages" }
 warpgate_api = { version = "0.1.3", path = "../warpgate-api" }
 anyhow = "1.0.75"
 semver = { workspace = true, features = ["serde"] }
